checkpoint: bad optimizations
Dramatically smaller model that looks to get the solution quicker
than previous models but takes longer to prove it. While working
on this it was noted that the order of vars (by name) made a notable
difference in runtime. Further testing is needed to see if obtaining
the optimal result is consistently earlier. If so, it might be worth
running two models, one as a 'pre-solver' and one as proff.

- removed lodging nodes and made town -> sink flow the control
 for setting lodging x vars.